moved shared include files (public ones) over to new folder structure and updated so shared and shared tests build again
