+++ Shared/Shared.vcxproj
@@ -91,7 +91,7 @@
       <LanguageStandard>stdcpplatest</LanguageStandard>
       <PrecompiledHeader>Create</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>C:\src\DebugMonitor\3rdParty\inc;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)\3rdParty\inc;$(SolutionDir)\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
@@ -107,7 +107,7 @@
       <LanguageStandard>stdcpplatest</LanguageStandard>
       <PrecompiledHeader>Create</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>C:\src\DebugMonitor\3rdParty\inc;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)\3rdParty\inc;$(SolutionDir)\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
@@ -125,7 +125,7 @@
       <LanguageStandard>stdcpplatest</LanguageStandard>
       <PrecompiledHeader>Create</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>C:\src\DebugMonitor\3rdParty\inc;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)\3rdParty\inc;$(SolutionDir)\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
@@ -145,7 +145,7 @@
       <LanguageStandard>stdcpplatest</LanguageStandard>
       <PrecompiledHeader>Create</PrecompiledHeader>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
-      <AdditionalIncludeDirectories>C:\src\DebugMonitor\3rdParty\inc;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)\3rdParty\inc;$(SolutionDir)\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
@@ -155,25 +155,25 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="bad_owner_access.h" />
-    <ClInclude Include="collection.h" />
-    <ClInclude Include="command_result.h" />
-    <ClInclude Include="data_member.h" />
-    <ClInclude Include="environment_repository.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\bad_owner_access.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\collection.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\command_result.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\data_member.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\environment_repository.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\process_service.h" />
     <ClInclude Include="environment_repository_impl.h" />
-    <ClInclude Include="export.h" />
-    <ClInclude Include="file_service.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\shared_export.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\file_service.h" />
     <ClInclude Include="file_service_impl.h" />
-    <ClInclude Include="invalid_handle.h" />
-    <ClInclude Include="null_handle.h" />
-    <ClInclude Include="process.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\invalid_handle.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\null_handle.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\process.h" />
     <ClInclude Include="process_impl.h" />
-    <ClInclude Include="process_service.h" />
     <ClInclude Include="process_service_impl.h" />
-    <ClInclude Include="not_found_exception.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\not_found_exception.h" />
     <ClInclude Include="pch.h" />
-    <ClInclude Include="string_extensions.h" />
-    <ClInclude Include="unique_handle.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\string_extensions.h" />
+    <ClInclude Include="$(SolutionDir)\include\shared\unique_handle.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="environment_repository_impl.cpp" />
