+++ Tests/SharedGoogleTests/file_service.cpp
@@ -12,7 +12,7 @@
 // 
 
 #include "pch.h"
-#include <file_service_impl.h>
+#include "../Shared/file_service_impl.h"
 #include "Common.h"
 
 using std::filesystem::directory_entry;
