+++ include/shared/bad_owner_access.h
