+++ include/shared/collection.h
