+++ include/shared/null_handle.h
