+++ include/shared/string_extensions.h
