+++ include/shared/unique_handle.h
